pyproject.toml
[tool.poetry]
name = "workspace"
version = "0.1.0"
description = ""
authors = ["Nathali Cardoso <nathalicardoso998@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.43.0"
pandas = "^2.2.3"
sqlalchemy = "^2.0.38"
psycopg2-binary = "^2.9.10"
plotly = "^6.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
